scripts/requirements.txt
# OpenAI SDK (gpt-image-2 support)
openai>=1.55.0

# Async HTTP (the OpenAI SDK already uses httpx; aiohttp is here for any
# extra tooling that might want a separate async client)
aiohttp>=3.9.0

# YAML templates and task lists
PyYAML>=6.0.1

# Image processing (slicing, alignment, sprite-sheet packing)
Pillow>=10.4.0

# Transparent-background removal (gpt-image-2 does not support transparent
# output, so we cut it out locally). rembg is large and downloads an ONNX
# model on first run.
rembg>=2.0.59
onnxruntime>=1.18.0

# .env file loader
python-dotenv>=1.0.1

# CLI prettifier (progress bars, tables)
rich>=13.7.0
